+++ src/main/java/Chapter07/Item45/smlee/Client.java
@@ -1,4 +1,4 @@
-package Chapter07.Item45;
+package Chapter07.Item45.smlee;
 
 import java.util.Arrays;
 import java.util.stream.DoubleStream;
