+++ src/main/java/Chapter07/Item46/smlee/Client.java
@@ -1,4 +1,4 @@
-package Chapter07.Item46;
+package Chapter07.Item46.smlee;
 
 import java.io.File;
 import java.io.IOException;
